requirements.txt
python-dotenv
djangorestframework 
drf-yasg